+++ packages/pangea-multipass-llama-index/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Slack channel's reader and processor.
 - `account_id` support on JiraProcessor.
 - Check user email permissions with admin credentials in GDriveProcessor.
+- Check username permissions with admin token in GithubProcessor.
 
 ## 0.1.0 - 2024-12-24